make employee table into its own component
