+++ react/src/components/dashboard/Dashboard.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import clsx from 'clsx';
-import { makeStyles, withStyles } from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Drawer from '@material-ui/core/Drawer';
 import AppBar from '@material-ui/core/AppBar';
@@ -14,17 +14,10 @@ import Grid from '@material-ui/core/Grid';
 import Paper from '@material-ui/core/Paper';
 import MenuIcon from '@material-ui/icons/Menu';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import { Link as RouterLink } from 'react-router-dom';
 import Button from '@material-ui/core/Button';
 import Box from '@material-ui/core/Box';
 import { mainListItems, secondaryListItems } from './listItems';
-import Modal from './Modal';
-
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
+import EmployeeTable from './EmployeeTable';
 
 const drawerWidth = 240;
 
@@ -110,88 +103,6 @@ const useStyles = makeStyles(theme => ({
    }
 }));
 
-//Employee table styles
-const StyledTableCell = withStyles(theme => ({
-   head: {
-      backgroundColor: theme.palette.common.black,
-      color: theme.palette.common.white
-   },
-   body: {
-      fontSize: 14
-   }
-}))(TableCell);
-
-const StyledTableRow = withStyles(theme => ({
-   root: {
-      '&:nth-of-type(odd)': {
-         backgroundColor: theme.palette.background.default
-      }
-   }
-}))(TableRow);
-
-function createData(name, calories, fat, carbs, protein) {
-   return { name, calories, fat, carbs, protein };
-}
-
-const rows = [
-   createData(
-      'John Snow',
-      'The King in the North',
-      80000,
-      <div>
-         <button>Edit</button>
-         <button>Delete</button>
-      </div>
-   ),
-   createData(
-      'Arya Stark',
-      'The Faceless Woman',
-      75000,
-      <div>
-         <button>Edit</button>
-         <button>Delete</button>
-      </div>
-   ),
-   createData(
-      'Daenerys Targaryen',
-      'The Mad Queen',
-      90000,
-      <div>
-         <button>Edit</button>
-         <button>Delete</button>
-      </div>
-   ),
-   createData(
-      'Sansa Stark',
-      'the Queen of WinterFell',
-      65000,
-      <div>
-         <button>Edit</button>
-         <button>Delete</button>
-      </div>
-   ),
-   createData(
-      'Sandor Clegane',
-      'The Hound',
-      55000,
-      <div>
-         <button>Edit</button>
-         <button>Delete</button>
-      </div>
-   )
-];
-
-const useStylesTable = makeStyles(theme => ({
-   root: {
-      width: '100%',
-      marginTop: theme.spacing(3),
-      overflowX: 'auto'
-   },
-   table: {
-      minWidth: 700
-   }
-}));
-
 export default function Dashboard() {
    const classes = useStyles();
    const [open, setOpen] = React.useState(true);
@@ -202,8 +113,6 @@ export default function Dashboard() {
       setOpen(false);
    };
    const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
-
-   const tableClasses = useStylesTable();
 
    return (
       <div className={classes.root}>
@@ -273,47 +182,7 @@ export default function Dashboard() {
                      <Paper>
                         {/* Employee table */}
                         <Box my={2}>
-                           <Paper className={classes.root}>
-                              <Table className={classes.table}>
-                                 <TableHead>
-                                    <TableRow>
-                                       <StyledTableCell>
-                                          Employee Name
-                                       </StyledTableCell>
-                                       <StyledTableCell align="right">
-                                          Job Title
-                                       </StyledTableCell>
-                                       <StyledTableCell align="right">
-                                          Salary
-                                       </StyledTableCell>
-                                       <StyledTableCell align="right">
-                                          Edit/Delete
-                                       </StyledTableCell>
-                                    </TableRow>
-                                 </TableHead>
-                                 <TableBody>
-                                    {rows.map(row => (
-                                       <StyledTableRow key={row.name}>
-                                          <StyledTableCell
-                                             component="th"
-                                             scope="row"
-                                          >
-                                             {row.name}
-                                          </StyledTableCell>
-                                          <StyledTableCell align="right">
-                                             {row.calories}
-                                          </StyledTableCell>
-                                          <StyledTableCell align="right">
-                                             {row.fat}
-                                          </StyledTableCell>
-                                          <StyledTableCell align="right">
-                                             {row.carbs}
-                                          </StyledTableCell>
-                                       </StyledTableRow>
-                                    ))}
-                                 </TableBody>
-                              </Table>
-                           </Paper>
+                           <EmployeeTable />
                         </Box>
                      </Paper>
                   </Grid>
